pyproject.toml
[tool.poetry]
name = "script-license-manager"
version = "0.1.0"
description = ""
authors = ["inna-tuzhikova <inna-tuzhikova@github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
drf-yasg = "^1.21.7"
django-filter = "^23.5"
jsonschema = "^4.21.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
isort = "^5.13.2"
flake8-pyproject = "^1.2.3"


[tool.poetry.group.prod.dependencies]
uwsgi = "^2.0.24"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 80
multi_line_output = 3
include_trailing_comma = true
use_parentheses = true
ensure_newline_before_comments = true
skip = ["scripts/migrations"]

[tool.flake8]
ignore = ["W503", "F811", "E722"]
max-line-length = 80
exclude = ["script_license_manager/settings.py", "scripts/migrations"]
